requirements.txt
absl-py==0.10.0
appdirs==1.4.4
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
distlib==0.3.1
filelock==3.0.12
gast==0.3.3
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
importlib-resources==3.0.0
Keras-Preprocessing==1.1.2
Markdown==3.2.2
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-gpu==2.3.0
tensorflow-gpu-estimator==2.3.0
termcolor==1.1.0
urllib3==1.25.10
virtualenv==20.0.30
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
